libogc/si: Expose SI_SetXY() and SI_SetSamplingRate() in the public header
These were likely intended to be exposed, but were missed or forgotten about. If anything,
SI_SetSamplingRate() was very likely intended to be exposed, as it sets the internal
sampling_rate variable that SI_RefreshSamplingRate() uses. Without it, any time
SI_RefreshSamplingRate() is called, it would reset the sampling rate to zero, which is likely
not always desirable.