video: add RGB 480p progressive mode
This mode was used in the Wii arcade board (RVA-001). Detecting it
for use with VIDEO_GetPreferredMode() would require some logic
involving reading /title/00000001/00000002/data/RVA.txt since that
is the main method used to detect the board instead of SC.
Due to the relative clumsiness involved I decided to leave it out
for now.