gu: add a function to multiply 4x4 matrices (#172)
Add guMtx44Concat() in both C and PS variants. The PS variant is more
than 3 times faster that the C one; it's written in a separate file (and
not in gu_psasm.S) because it uses a different naming of the matrix
registers.

These functions are especially useful when porting programs written for
OpenGL, which uses 4x4 matrices.