gx: make GX_SetTevIndTile() more readable
The old code suffered from some obfuscation due to how it was reverse
engineered:
- The 0x43300000 magix number is the binary representation of the upper
  32 bits of a double floating point number havint the exponent set to
  1075, which makes it so that the least significant bit of the lower 32
  bit word maps exactly to the unit 1.
- 4503599627370496.0F is 1 << 52, which in floating point representation
  is 0x43300000 00000000.
- 0.00097656250F is "1.0 / (1 << 10)"

In other words: multiplying by 0.00097656250F is equivalent to dividing
by 2^10, and this is the actual operation that the old code was
fulfilling. The subtraction between the double floating point numbers is
just an optimisation to convert the integer parameter into a float, so
we can safely ignore it and leave it all to the compiler.