Video: add support for RGB/VGA output (#167)
* video: refactor magic register writes into named functions

* video: add generic RGB mode flag and allow re-configuring encoder

Previously, the encoder was configured only once during VIDEO_Init(),
and would not be re-configured when VIDEO_Configure(mode) was called.
Since we can switch from a non-RGB to an RGB mode, we need to be able
to reconfigure the encoder at configure-time. This commit extends the
mode-switching mechanism to also allow scheduling for encoder setting
changes at run-time.

Moreover, since there are more RGB video modes than EURGB60, we extend
the mode struct with a flag to indicate whether the mode is RGB or not
or not for easy internal use. Adding it to the VI_TVMODE enumeration
was also considered, but decided against since it would break the
VIDEO_GetCurrentTvMode() ABI for EURGB60 modes.

* video: fix encoder setup for RGB video modes

* video: add RGB 480p progressive mode

This mode was used in the Wii arcade board (RVA-001). Detecting it
for use with VIDEO_GetPreferredMode() would require some logic
involving reading /title/00000001/00000002/data/RVA.txt since that
is the main method used to detect the board instead of SC.
Due to the relative clumsiness involved I decided to leave it out
for now.